Created Authorization Class that have a isOwner method to not allow a user delete or modify a listing if not authorized.

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -22,7 +24,9 @@ class ListingController
    */
   public function index()
   {
-    $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+    $listings = $this->db
+      ->query('SELECT * FROM listings ORDER BY created_at DESC')
+      ->fetchAll();
 
     loadView('listings/index', [
       'listings' => $listings,
@@ -91,7 +95,7 @@ class ListingController
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     $newListingData = array_map('sanitize', $newListingData);
 
@@ -169,11 +173,19 @@ class ListingController
       ->query('SELECT * FROM listings WHERE id = :id', $params)
       ->fetch();
 
+    // Check if listing exists
     if (!$listing) {
       ErrorController::notFound('Listing not found.');
     }
 
-    // make DELETE command if $listing exists
+    // Authorization
+    if (!Authorization::isOwner($listing->user_id)) {
+      $_SESSION['error_message'] =
+        'You are not authorized to delete this listing.';
+      return redirect('/listings/' . $listing->id);
+    }
+
+    // make DELETE in database if $listing exists
     $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
     // Set flash message
